Fixed/refactored the ModelProperties class to rely on information from the ParmEd Structure
